backend/requirements.txt
# FastAPI & Server
# ✅ PHASE 2.1: Updated to latest stable versions
fastapi==0.119.0  # was 0.104.1 (+15 minor versions)
uvicorn[standard]==0.34.0  # was 0.24.0 (security + perf)
python-multipart==0.0.6

# Database
sqlalchemy==2.0.44  # was 2.0.23 (21 patch versions, security fixes)
alembic==1.14.1  # was 1.12.1 (bug fixes)
asyncpg==0.30.0  # was 0.29.0
psycopg2-binary==2.9.10  # was 2.9.9

# Pydantic
pydantic[email]==2.12.3  # was 2.5.0 (+7 minor versions, validation improvements)
pydantic-settings==2.8.1  # was 2.1.0
email-validator==2.2.0  # was 2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Redis & Celery
redis==5.0.1
celery==5.3.4

# Rate Limiting
slowapi==0.1.9

# Paddle (Payments - PAS Stripe)
requests==2.31.0
httpx==0.25.2

# Crypto Exchange APIs
ccxt==4.5.10

# Blockchain & Web3
web3==6.15.1
eth-abi==4.2.1
base58==2.1.1
moralis==0.1.49  # Moralis Web3 API for DeFi parsing (updated to fix typing-extensions conflict)

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0

# AI & LLM
anthropic==0.71.0  # Claude AI for chat assistant

# Web Scraping & Data Fetching
beautifulsoup4==4.12.2
lxml==5.1.0
pdfplumber==0.10.3
PyPDF2==3.0.1

# PDF Generation
weasyprint==60.2
pydyf==0.10.0

# Monitoring & Error Tracking
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# DeFi Audit Enhancements
flower==2.0.1  # Celery monitoring UI
pandas==2.1.4  # Data processing & CSV
openpyxl==3.1.2  # Excel export
reportlab==4.0.7  # Advanced PDF generation
matplotlib==3.8.2  # Charts in reports
sendgrid==6.11.0  # Email notifications
twilio==8.11.1  # SMS alerts
aiofiles==23.2.1  # Async file operations
